2lab/requirements.txt
SQLAlchemy~=1.4.34
fastapi~=0.115.12
pydantic~=2.11.2
passlib~=1.7.4
typing~=3.7.4.3
PyJWT~=2.10.1
